fix some styles depend of device
